feature: Switching over to web socket.
Long story short-  it looks like I can't get electron running reliably on
raspbian vis a vis the pi zero.

Realizing that between this and my inability to build serialport on
windows for electron, making this a standalone app was taking too much
time. I decided to take the easier router and break the app down into a
ws/express server that serves up a local web page, where the backend
handles all of the logic as per the original electron server's main
process.

Also, a few typos and changes and small bugs fixed as I was
transferring/running/testing